booking_service/requirements.txt
Django>=4.2,<6.1
djangorestframework
djangorestframework-simplejwt
django-environ
psycopg2-binary
django-filter
django-cors-headers
